requirements.txt
os
json
Pillow
trimesh
io